src/test/java/com/playground/spring/controller/RandomControllerTest.java
package com.playground.spring.controller;

import com.fasterxml.jackson.databind.ObjectMapper;
import com.playground.spring.dto.GetRandomResponse;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.http.MediaType;
import org.springframework.test.context.ContextConfiguration;
import org.springframework.test.web.servlet.MockMvc;

import static org.junit.jupiter.api.Assertions.assertNotNull;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

@WebMvcTest
@AutoConfigureMockMvc
@ContextConfiguration(classes = {RandomController.class})
class RandomControllerTest {

    @Autowired
    private MockMvc mvc;

    @Autowired
    private ObjectMapper objectMapper;

    @Test
    public void testGetRandom() throws Exception {
        String contentAsString = mvc.perform(get("/random")
                .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isOk())
                .andReturn()
                .getResponse()
                .getContentAsString();

        GetRandomResponse getRandomResponse = objectMapper.readValue(contentAsString, GetRandomResponse.class);

        assertNotNull(getRandomResponse.getRandom());
    }

}